refactor: move and modify yaml loading routine
body: This commit is to revert the yaml loading routine to
what it was prior to the unit test addition, however it is moved
to it's own method so that the unit test can substitute it's own
version of 'load_yaml_file'.  The logger is added to the unit test
as well as the new function'load_yaml_file'.
