+++ src/eva/tests/test_obs_correlation_scatter.py
@@ -4,6 +4,11 @@ import pytest
 from eva import eva_base
 from unittest.mock import patch
 
+# local imports
+from eva.utilities.logger import Logger
+from eva.tests import helpers
+
+
 EVA_TESTS_DIR = pathlib.Path(__file__).parent.resolve()
 PYTEST_ENVVARS = {
     'EVA_TESTS_DIR': str(EVA_TESTS_DIR)
@@ -11,6 +16,9 @@ PYTEST_ENVVARS = {
 
 OBS_CORRELATION_SCATTER_YAML = os.path.join(
     EVA_TESTS_DIR, 'config/ObsCorrelationScatterDriver.yaml')
+
+# redirect load_yaml_file from eva_base to helpers
+eva_base.load_yaml_file = helpers.load_yaml_file
 
 
 def test_obs_correlation_scatter():
